requirements.txt
loremipsum
numpy
Pillow
pylint
requests
six
